[CINN] Remove split_transform for Reduce
