+++ paddle/cinn/hlir/framework/pir/trivial_op_util.h
@@ -169,8 +169,6 @@ extern ExprSetFinder ScheduleBlockRealizeIsNotInit;
 
 extern ExprSetFinder ScheduleBlockRealizeIsInit;
 
-extern ExprSetFinder ScheduleBlockRealizeIsSplitTransform;
-
 extern ExprSetFinder IsFor;
 
 extern ExprSetFinder ChildScheduleBlocks;
