+++ paddle/cinn/ir/group_schedule/config/group_tile_util.h
@@ -18,6 +18,38 @@
 namespace cinn {
 namespace ir {
 
+/**
+ * Get the strides of loops according to the largest input of Reduce.
+ *
+ * For example, in the following compute body of Reduce:
+ *   for (i, 0, 8):
+ *     for (j, 0, 16):
+ *       for (k, 0, 24):
+ *         for (m, 0, 32):
+ *           var[i] = var[i] + var_1[k, i, j, m] * var_2[i, k, j]
+ *
+ * There are two inputs, whose sizes are:
+ *    var_1[k, i, j, m]: 24*8*16*32 = 98304
+ *    var_2[i, k, j]: 8*24*16 = 3072
+ * We can see that the largest input is var_1[k, i, j, m].
+ *
+ * Therefore, the strides of loops (i, j, k and m) according to the largest
+ * input (var_1[k, i, j, m]) are:
+ *    i:   16*32 = 512
+ *    j:      32 = 32
+ *    k: 8*16*32 = 4096
+ *    m:       1 = 1
+ *
+ * Limitations:
+ * 1) If there are multiple inputs of the same size, we simply choose the first
+ *    visited input.
+ * 2) If the input's size contains symbols (e.g. S0, S1), we just replace each
+ *    symbol with 32 so as to get a constant stride.
+ *
+ * @param reduce_compute_body The root For node of Reduce.
+ */
+std::vector<int64_t> GetLoopStrides(const ir::Expr& reduce_compute_body);
+
 // Check whether we can apply grid reduce in this group.
 // We can apply grid reduce if there is no reduce-then-broadcast dependency
 // in this group.
