+++ paddle/cinn/operator_fusion/fusion_tracker/expr_utils.h
@@ -30,16 +30,6 @@ struct FusibleOp2Expr {
   }
 };
 
-struct GetSplitedExprFromFusionOp {
-  std::vector<ir::Expr> operator()(const TrivialOp& op) {
-    return {op.GetFuncBody()};
-  }
-  std::vector<ir::Expr> operator()(const ReduceOp& op) {
-    const auto& t_r = SplitReduceOp(op);
-    return {t_r.first.GetFuncBody(), t_r.second.GetFuncBody()};
-  }
-};
-
 struct ApplyItersTransform {
   explicit ApplyItersTransform(const ir::Expr& expr,
                                const ir::Expr& aligned_expr)
